renaming `eccs` folder to `ecc`
